Refactor Slack service to separate modules
