+++ app/controllers/slack_controller.rb
@@ -6,46 +6,46 @@ class SlackController < ApplicationController
 
   def give_kudos
     begin
-      SlackService.create_post(params[:text], params[:team_id], params[:user_id])
+      Slack::SlackService.create_post(params[:text], params[:team_id], params[:user_id])
       render json: {text: 'kudos are on the way!'}
-    rescue SlackService::InvalidCommand => e
+    rescue Slack::SlackService::InvalidCommand => e
       return render json: {text: "That didn't quite work, #{e} \n The format is: /kudo @someone @someone.else [amount] for [reason]"}
     end
   end
 
   def team_auth_callback
     begin
-      SlackService.add_to_workspace(params[:code], params[:team_id])
+      Slack::SlackService.add_to_workspace(params[:code], params[:team_id])
       redirect_to Settings.slack_team_connect_success_url
-    rescue SlackService::InvalidRequest => e
+    rescue Slack::SlackService::InvalidRequest => e
       render json: {text: "That didn't quite work, #{e}"}
     end
   end
 
   def user_auth_callback
     begin
-      SlackService.connect_account(params[:code], params[:user_id])
+      Slack::Slack::SlackService.connect_account(params[:code], params[:user_id])
       redirect_to Settings.slack_user_connect_success_url
-    rescue SlackService::InvalidRequest => e
+    rescue Slack::Slack::SlackService::InvalidRequest => e
       render json: {text: "That didn't quite work, #{e}"}
     end
   end
 
   def guidelines
     begin
-      payload = SlackService.list_guidelines(params[:team_id])
+      payload = Slack::SlackService.list_guidelines(params[:team_id])
       render json: {blocks: payload}
-    rescue SlackService::InvalidCommand => e
+    rescue Slack::SlackService::InvalidCommand => e
       render json: {text: "That didn't quite work, #{e}"}
     end
   end
 
   def auth_team
-    redirect_to SlackService.get_team_oauth_url(params[:team_id])
+    redirect_to Slack::SlackService.get_team_oauth_url(params[:team_id])
   end
 
   def auth_user
-    redirect_to SlackService.get_user_oauth_url(params[:user_id])
+    redirect_to Slack::Slack::SlackService.get_user_oauth_url(params[:user_id])
   end
 
   def event
