+++ app/services/slack/slack_service.rb
@@ -1,0 +1,13 @@
+module Slack
+  class SlackService
+    extend Auth
+    extend Command
+    extend Event
+    extend Message
+    extend Support
+
+    class InvalidRequest < RuntimeError; end
+
+    class InvalidCommand < RuntimeError; end
+  end
+end
