+++ app/workers/slack_worker.rb
@@ -5,9 +5,9 @@ class SlackWorker
   def perform(team_id, event)
     case event['type']
     when 'reaction_added'
-      SlackService.reaction_added(team_id, event)
+      Slack::SlackService.reaction_added(team_id, event)
     when 'reaction_removed'
-      SlackService.reaction_removed(team_id, event)
+      Slack::SlackService.reaction_removed(team_id, event)
     else
       raise RuntimeError.new('Unsupported event')
     end
